+++ src/client/cypress/e2e/mainPage/copyBorehole.cy.js
@@ -17,9 +17,7 @@ describe("Test copying of boreholes", () => {
     cy.get(".MuiDataGrid-checkboxInput").eq(1).scrollIntoView();
     cy.get(".MuiDataGrid-checkboxInput").eq(1).click();
     cy.contains("button", "Create a copy").click();
-
-    cy.get(".MuiDialogContentText-root").should("contain.text", "Select a workgroup to create a copy.");
-    cy.get(".MuiButton-contained").contains("Create a copy").click();
+    handlePrompt("Select a workgroup to create a copy.", "copy");
     cy.wait("@borehole_copy");
     cy.wait("@borehole_get");
 
@@ -41,9 +39,10 @@ describe("Test copying of boreholes", () => {
     cy.get(".MuiDataGrid-checkboxInput").eq(1).click();
     cy.contains("button", "Create a copy").click();
 
-    cy.get(".MuiDialogContentText-root").should("contain.text", "Select a workgroup to create a copy.");
-    setSelect("workgroup", 1);
-    cy.get(".MuiButton-contained").contains("Create a copy").click();
+    cy.get('[data-cy="prompt"]').should("be.visible");
+    cy.get('[data-cy="prompt"]').contains("Select a workgroup to create a copy.");
+    setSelect("workgroup", 1, null, "prompt");
+    cy.get('[data-cy="prompt"]').find(`[data-cy="copy-button"]`).click();
     cy.wait("@borehole_copy");
     cy.wait("@borehole_get");