+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Boreholes were always copied to the default workgroup instead of the selected workgroup.
 - _view-sync_ failed if the source database had a default workgroup named `Default`.
 - Depths with thousand separators could not be input in lithostratigraphy and chronostratigraphy.
+- Fixed a bug where copy boreholes dialog could not always be closed.
 
 ## v2.1.1162 - 2025-04-10